Request object
Request object can be used straight or through the usage of other functions of
this library. Request object is a convenience object which gives the soap
request template in javascript, which is then parsed to xml. This abstraction
layer also provides some useful setters and getters to save some typing.